pyproject.toml
[project]
name = "agnitio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.136.0",
    "google-genai>=1.73.1",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy==2.0.49",
    "uvicorn>=0.44.0",
]
extra = '''
[tool.pyright]
pythonVersion = \"3.11\"

[tool.ruff]
target-version = \"py311\"
'''
